fixed match creation so it doesn't do it more than once

+++ awardbjj/app/controllers/brackets_and_matches_controller.rb
@@ -1,19 +1,28 @@
 class BracketsAndMatchesController < ApplicationController
 	load_and_authorize_resource :event
 
-	# TODO: do only once. If done twice, override the matches
 	def create
 		# for the future: add three person comeback, single elimination with bronze, ...
-		# add an if else to check what create_matches returns
-		@event.create_matches
-		redirect_to event_brackets_path(@event), notice: 'Matches were successfully created.'
+		if @event.matches.empty?
+			if @event.create_matches
+				redirect_to event_brackets_path(@event), notice: 'Matches were successfully created.'
+			else
+				redirect_to event_brackets_path(@event), notice: 'Matches were not successfully created.'
+			end
+		else
+			redirect_to event_brackets_path(@event), alert: 'Matches have already been created.'
+		end
 	end
 
 	def destroy
-		@event.brackets.each do |bracket|
-			bracket.matches.delete_all # using delete_all, because it does not trigger callbacks
+		if @event.matches.empty?
+			# TODO: don't reload full page, use flash.now[:notice] = 'There are no matches to destroy.'
+			redirect_to event_brackets_path(@event), notice: 'There are no matches to destroy.'
+		else
+			@event.brackets.each do |bracket|
+				bracket.matches.delete_all # using delete_all, because it does not trigger callbacks
+			end
+			redirect_to event_brackets_path(@event), notice: 'Matches were successfully destroyed.'
 		end
-    
-		redirect_to event_brackets_path(@event), notice: 'Matches were successfully destroyed.'
 	end
 end
